kafka/consumer/requirements.txt
psutil==5.9.4
kafka-python==2.0.2
psycopg2-binary==2.9.10
channels==4.2.0
asgiref==3.8.1
redis==5.2.1
django==5.1.3
daphne==4.1.2
channels_redis==4.1.0